task updated by dev1

+++ Proj01/src/main/java/com/nt/main/App.java
@@ -5,7 +5,7 @@ package com.nt.main;
  */
 public class App {
 	
-	public int Sum(int a,int b) {
+	public double Sum(double a,double b) {
 		return a+b;
 	}
     public static void main(String[] args) {
